requirements.txt
flask==3.0.3
flask-cors==4.0.1
scikit-learn==1.4.2
pandas==2.2.2
numpy==1.26.4
joblib==1.4.2
google-generativeai==0.7.2
gunicorn==22.0.0
